+++ services/src/main/java/io/scalecube/services/RemoteServiceInstance.java
@@ -1,5 +1,7 @@
 package io.scalecube.services;
 
+import static com.google.common.base.Preconditions.checkArgument;
+
 import io.scalecube.transport.Address;
 import io.scalecube.transport.Message;
 
@@ -12,13 +14,15 @@ import rx.Subscription;
 import rx.subjects.PublishSubject;
 import rx.subjects.Subject;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicReference;
 
 public class RemoteServiceInstance implements ServiceInstance {
- 
+
   private static final Logger LOGGER = LoggerFactory.getLogger(RemoteServiceInstance.class);
 
   private final Address address;
@@ -42,7 +46,7 @@ public class RemoteServiceInstance implements ServiceInstance {
       Map<String, String> tags) {
 
     this.serviceName = serviceReference.serviceName();
-    this.methods = serviceReference.methods();
+    this.methods = Collections.unmodifiableSet(serviceReference.methods());
     this.address = serviceReference.address();
     this.memberId = serviceReference.memberId();
     this.tags = tags;
@@ -65,17 +69,13 @@ public class RemoteServiceInstance implements ServiceInstance {
   public Observable<Message> listen(final Message request) {
 
     final String cid = request.correlationId();
-    
+
     AtomicReference<Subscription> subscription = new AtomicReference<>();
     Observable<Message> observer = transportObservable.doOnUnsubscribe(() -> {
       Message unsubscribeRequest = Messages.asUnsubscribeRequest(cid);
       LOGGER.info("sending remote unsubscribed event: {}", unsubscribeRequest);
       subscription.get().unsubscribe();
-      sendRemote(unsubscribeRequest).whenComplete((success, error) -> {
-        if (error != null) {
-          LOGGER.error("Failed sending remote unsubscribed event: {} {}", unsubscribeRequest, error);
-        }
-      });
+      sendRemote(unsubscribeRequest);
     });
 
     Subscription sub = observer
@@ -83,15 +83,9 @@ public class RemoteServiceInstance implements ServiceInstance {
         .subscribe(onNext -> {
           serviceResponses.onNext(onNext);
         });
-    
+
     subscription.set(sub);
-
-    sendRemote(request).whenComplete((success, error) -> {
-      if (error != null) {
-        LOGGER.error("Failed sending remote subscribe request: {} {}", request, error);
-      }
-    });
-
+    sendRemote(request);
     return serviceResponses;
   }
 
@@ -111,31 +105,22 @@ public class RemoteServiceInstance implements ServiceInstance {
 
   @Override
   public CompletableFuture<Message> invoke(Message request) {
-
     Messages.validate().serviceRequest(request);
-
     CompletableFuture<Message> result = new CompletableFuture<Message>();
-
-    futureInvoke(request)
-        .whenComplete((success, error) -> {
-          if (error == null) {
-            result.complete(Message.builder().data("remote send completed").build());
-          } else {
-            LOGGER.error("Failed to send request {} to target address {}", request, address);
-            result.completeExceptionally(error);
-          }
-        });
-
+    
+    try {
+      sendRemote(request);
+      result.complete(Message.builder().data("remote send completed").build());
+    } catch (Exception error) {
+      LOGGER.error("Failed to send request {} to target address {}", request, address);
+      result.completeExceptionally(error);
+    }
     return result;
   }
 
-  private CompletableFuture<Void> futureInvoke(final Message request) {
-    return sendRemote(request);
-  }
-
-  private CompletableFuture<Void> sendRemote(Message request) {
+  private void sendRemote(Message request) {
     LOGGER.debug("cid [{}] send remote service request message {}", request.correlationId(), request);
-    return this.sender.send(address, request);
+    this.sender.send(address, request);
   }
 
   @Override
@@ -163,13 +148,25 @@ public class RemoteServiceInstance implements ServiceInstance {
   }
 
   @Override
-  public boolean hasMethod(String methodName) {
+  public boolean methodExists(String methodName) {
     return methods.contains(methodName);
   }
-  
+
+  @Override
+  public void checkMethodExists(String methodName) {
+    checkArgument(this.methodExists(methodName), "instance has no such requested method");
+  }
+
+  @Override
+  public Collection<String> methods() {
+    return methods;
+  }
+
   @Override
   public String toString() {
     return "RemoteServiceInstance [serviceName=" + serviceName + ", address=" + address + ", memberId=" + memberId
         + ", methods=" + methods + ", tags=" + tags + "]";
   }
+
+
 }